split grpc server implementation into multiple files
